tests/wpt/web-platform-tests/css/css-animations/parsing/animation-range-start-valid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-range">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
// https://drafts.csswg.org/scroll-animations-1/#view-timelines-ranges
test_valid_value("animation-range-start", "auto");
test_valid_value("animation-range-start", "cover 0%");
test_valid_value("animation-range-start", "cover 100%");
test_valid_value("animation-range-start", "cover 120%");
test_valid_value("animation-range-start", "cover 42%");
test_valid_value("animation-range-start", "cover -42%");
test_valid_value("animation-range-start", "contain 42%");
test_valid_value("animation-range-start", "exit 42%");
test_valid_value("animation-range-start", "exit 1%, cover 2%, contain 100%");
test_valid_value("animation-range-start", "enter 42px");
test_valid_value("animation-range-start", "contain calc(10px + 10%)");
test_valid_value("animation-range-start", "enter 1em");
test_valid_value("animation-range-start", "exit calc(1em + 10px)");

// There's an open issue in the spec about "enter" vs "entry".
//
// https://drafts.csswg.org/scroll-animations-1/#valdef-animation-timeline-range-entry
test_valid_value("animation-range-start", "enter 42%");
</script>
